Blurbify, TrilogyCO -> BuggedChargeCO
